Add basic implementation of scattered data interpolation
This script implements scattered data interpolation, which means that given a number of data points -- that are scattered through some space -- we can interpolate between them. Specifically, this type of interpolation using radial-basis functions paired with a Guassian kernel.

I plan to use this module to interpolate elements in the Lee algebra.